+++ Classes/XeePNMLoader.m
@@ -12,7 +12,7 @@
 +(BOOL)canOpenFile:(NSString *)name firstBlock:(NSData *)block attributes:(NSDictionary *)attributes;
 {
 	const unsigned char *head=[block bytes];
-	int len=[block length];
+	NSInteger len=[block length];
 
 	if(len>=3)
 	if(head[0]=='P')
